Post-Phase-10 audit fixes: branch-clone + int-overflow hardening
Mandatory post-phase audit (Rule 9) after Phase 10 + 10.5 completion.
0 critical / 0 high / ~15 actionable out of 1503 raw (~99% noise
floor). Verified by cppcheck, clang-tidy, semgrep, and gitleaks.

Real bug: audio_music_stream's dead if/else (both arms set the same
value) collapsed to a single assignment — bugprone-branch-clone.

Hardening: motion-matching feature-index math (kd_tree + motion_database,
13 sites) now widens operands to size_t before multiply, eliminating
int32 overflow risk on large motion databases. Fixes bugprone-misplaced-
widening-cast + bugprone-implicit-widening-of-multiplication-result
flagged across the two files.

Perf/style: clipNames() returns const-ref, ApplyTargets ctor arg is
const-ref, MAX_AUDIO_FRAMES uses ULL suffix, FilterTmpDir ctor uses
initializer list.

Build clean, 2661 tests pass.

+++ engine/animation/motion_database.cpp
@@ -57,7 +57,7 @@ void MotionDatabase::build(const FeatureSchema& schema,
     // Normalize the feature matrix in-place
     for (int f = 0; f < numFrames; ++f)
     {
-        m_normalizer.normalize(&m_features[static_cast<size_t>(f * m_numFeatures)],
+        m_normalizer.normalize(&m_features[static_cast<size_t>(f) * static_cast<size_t>(m_numFeatures)],
                                m_numFeatures);
     }
 
@@ -73,7 +73,8 @@ void MotionDatabase::build(const FeatureSchema& schema,
     {
         for (int d = 0; d < m_numFeatures; ++d)
         {
-            m_features[static_cast<size_t>(f * m_numFeatures + d)] *=
+            m_features[static_cast<size_t>(f) * static_cast<size_t>(m_numFeatures)
+                       + static_cast<size_t>(d)] *=
                 sqrtWeights[static_cast<size_t>(d)];
         }
     }
@@ -204,7 +205,7 @@ void MotionDatabase::sampleClip(int clipIdx, const AnimClipEntry& entry,
 void MotionDatabase::extractFeatures()
 {
     int numFrames = static_cast<int>(m_poses.size());
-    m_features.resize(static_cast<size_t>(numFrames * m_numFeatures), 0.0f);
+    m_features.resize(static_cast<size_t>(numFrames) * static_cast<size_t>(m_numFeatures), 0.0f);
 
     const auto& trajSamples = m_schema.getTrajectorySamples();
     int numTrajSamples = static_cast<int>(trajSamples.size());
@@ -271,7 +272,7 @@ void MotionDatabase::extractFeatures()
 
         FeatureExtractor::extract(m_schema, pose, rootPos, rootRotY,
                                   trajPositions.data(), trajDirections.data(),
-                                  &m_features[static_cast<size_t>(f * m_numFeatures)]);
+                                  &m_features[static_cast<size_t>(f) * static_cast<size_t>(m_numFeatures)]);
     }
 }
 
@@ -345,7 +346,8 @@ MotionSearchResult MotionDatabase::search(const float* query, uint32_t tagMask,
             for (int d = 0; d < m_numFeatures; ++d)
             {
                 float diff = normalizedQuery[static_cast<size_t>(d)]
-                           - m_features[static_cast<size_t>(i * m_numFeatures + d)];
+                           - m_features[static_cast<size_t>(i) * static_cast<size_t>(m_numFeatures)
+                                        + static_cast<size_t>(d)];
                 dist += diff * diff;
             }
 
@@ -485,7 +487,7 @@ void MotionDatabase::addMirroredFrames(
 
     for (int f = 0; f < numFrames; ++f)
     {
-        m_normalizer.normalize(&m_features[static_cast<size_t>(f * m_numFeatures)],
+        m_normalizer.normalize(&m_features[static_cast<size_t>(f) * static_cast<size_t>(m_numFeatures)],
                                m_numFeatures);
     }
 
@@ -494,7 +496,8 @@ void MotionDatabase::addMirroredFrames(
     {
         for (int d = 0; d < m_numFeatures; ++d)
         {
-            m_features[static_cast<size_t>(f * m_numFeatures + d)] *=
+            m_features[static_cast<size_t>(f) * static_cast<size_t>(m_numFeatures)
+                       + static_cast<size_t>(d)] *=
                 std::sqrt(weights[static_cast<size_t>(d)]);
         }
     }
